Break out poetry.go and pip.go
