+++ internal/backends/python/python.go
@@ -2,21 +2,16 @@
 package python
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"os"
-	"os/exec"
 	"regexp"
 	"strings"
 
 	"github.com/BurntSushi/toml"
 	"github.com/replit/upm/internal/api"
-	"github.com/replit/upm/internal/nix"
-	"github.com/replit/upm/internal/pkg"
 	"github.com/replit/upm/internal/util"
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 )
 
 var normalizationPattern = regexp.MustCompile(`[-_.]+`)
@@ -69,42 +64,6 @@ type pyprojectTOML struct {
 			Group           map[string]pyprojectTOMLGroup `json:"group"`
 		} `json:"poetry"`
 	} `json:"tool"`
-}
-
-// poetryLock represents the relevant parts of a poetry.lock file, in
-// TOML format.
-type poetryLock struct {
-	Package []struct {
-		Name    string `json:"name"`
-		Version string `json:"version"`
-	} `json:"package"`
-}
-
-func pipIsAvailable() bool {
-	_, err := exec.LookPath("pip")
-	return err == nil
-}
-
-func poetryIsAvailable() bool {
-	_, err := exec.LookPath("poetry")
-	return err == nil
-}
-
-// normalizeSpec returns the version string from a Poetry spec, or the
-// empty string. The Poetry spec may be either a string or a
-// map[string]interface{} with a "version" key that is a string. If
-// neither, then the empty string is returned.
-func normalizeSpec(spec interface{}) string {
-	switch spec := spec.(type) {
-	case string:
-		return spec
-	case map[string]interface{}:
-		switch spec := spec["version"].(type) {
-		case string:
-			return spec
-		}
-	}
-	return ""
 }
 
 func normalizePackageArgs(args []string) map[api.PkgName]api.PkgCoordinates {
@@ -208,45 +167,6 @@ func info(name api.PkgName) api.PkgInfo {
 	return info
 }
 
-func add(ctx context.Context, pkgs map[api.PkgName]api.PkgSpec, projectName string) {
-	//nolint:ineffassign,wastedassign,staticcheck
-	span, ctx := tracer.StartSpanFromContext(ctx, "poetry (init) add")
-	defer span.Finish()
-	// Initalize the specfile if it doesnt exist
-	if !util.Exists("pyproject.toml") {
-		cmd := []string{"poetry", "init", "--no-interaction"}
-
-		if projectName != "" {
-			cmd = append(cmd, "--name", projectName)
-		}
-
-		util.RunCmd(cmd)
-	}
-
-	cmd := []string{"poetry", "add"}
-	for name, spec := range pkgs {
-		name := string(name)
-		if found, ok := moduleToPypiPackageAliases[name]; ok {
-			delete(pkgs, api.PkgName(name))
-			name = found
-			pkgs[api.PkgName(name)] = api.PkgSpec(spec)
-		}
-		spec := string(spec)
-
-		// NB: this doesn't work if spec has
-		// spaces in it, because of a bug in
-		// Poetry that can't be worked around.
-		// It looks like that bug might be
-		// fixed in the 1.0 release though :/
-		if spec != "" {
-			cmd = append(cmd, name+" "+spec)
-		} else {
-			cmd = append(cmd, name)
-		}
-	}
-	util.RunCmd(cmd)
-}
-
 func searchPypi(query string) []api.PkgInfo {
 	// Normalize query before looking it up in the overide map
 	query = string(normalizePackageName(api.PkgName(query)))
@@ -275,136 +195,6 @@ func searchPypi(query string) []api.PkgInfo {
 	return filtered
 }
 
-// makePythonPoetryBackend returns a backend for invoking poetry, given an arg0 for invoking Python
-// (either a full path or just a name like "python3") to use when invoking Python.
-func makePythonPoetryBackend(python string) api.LanguageBackend {
-	return api.LanguageBackend{
-		Name:                 "python3-poetry",
-		Alias:                "python-python3-poetry",
-		Specfile:             "pyproject.toml",
-		IsSpecfileCompatible: verifyPoetrySpecfile,
-		Lockfile:             "poetry.lock",
-		IsAvailable:          poetryIsAvailable,
-		FilenamePatterns:     []string{"*.py"},
-		Quirks: api.QuirksAddRemoveAlsoLocks |
-			api.QuirksAddRemoveAlsoInstalls,
-		NormalizePackageArgs: normalizePackageArgs,
-		NormalizePackageName: normalizePackageName,
-		GetPackageDir: func() string {
-			// Check if we're already inside an activated
-			// virtualenv. If so, just use it.
-			if venv := os.Getenv("VIRTUAL_ENV"); venv != "" {
-				return venv
-			}
-
-			// Take PYTHONUSERBASE into consideration, if set
-			if userbase := os.Getenv("PYTHONUSERBASE"); userbase != "" {
-				return userbase
-			}
-
-			// Terminate early if we're running inside a repl.
-			// This will suppress the following poetry commands
-			// from showing up in the Packager pane.
-			if os.Getenv("REPL_HOME") != "" {
-				return ""
-			}
-
-			outputB, err := util.GetCmdOutputFallible([]string{
-				"poetry", "env", "list", "--full-path",
-			})
-			if err != nil {
-				// there's no virtualenv configured, so no package directory
-				return ""
-			}
-
-			var path string
-			for _, line := range strings.Split(strings.TrimSpace(string(outputB)), "\n") {
-				var isActive bool
-				path, isActive = strings.CutSuffix(line, " (Activated)")
-				if isActive {
-					break
-				}
-			}
-
-			return path
-		},
-		SortPackages: pkg.SortPrefixSuffix(normalizePackageName),
-
-		Search: searchPypi,
-		Info:   info,
-		Add:    add,
-		Remove: func(ctx context.Context, pkgs map[api.PkgName]bool) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "poetry remove")
-			defer span.Finish()
-			cmd := []string{"poetry", "remove"}
-			for name := range pkgs {
-				cmd = append(cmd, string(name))
-			}
-			util.RunCmd(cmd)
-		},
-		Lock: func(ctx context.Context) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "poetry lock")
-			defer span.Finish()
-			util.RunCmd([]string{"poetry", "lock", "--no-update"})
-		},
-		Install: func(ctx context.Context) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "poetry install")
-			defer span.Finish()
-			// Unfortunately, this doesn't necessarily uninstall
-			// packages that have been removed from the lockfile,
-			// which happens for example if 'poetry remove' is
-			// interrupted. See
-			// <https://github.com/sdispater/poetry/issues/648>.
-			util.RunCmd([]string{"poetry", "install"})
-		},
-		ListSpecfile: func(mergeAllGroups bool) map[api.PkgName]api.PkgSpec {
-			pkgs, err := listPoetrySpecfile(mergeAllGroups)
-			if err != nil {
-				util.DieIO("%s", err.Error())
-			}
-
-			return pkgs
-		},
-		ListLockfile: func() map[api.PkgName]api.PkgVersion {
-			var cfg poetryLock
-			if _, err := toml.DecodeFile("poetry.lock", &cfg); err != nil {
-				util.DieProtocol("%s", err.Error())
-			}
-			pkgs := map[api.PkgName]api.PkgVersion{}
-			for _, pkgObj := range cfg.Package {
-				name := api.PkgName(pkgObj.Name)
-				version := api.PkgVersion(pkgObj.Version)
-				pkgs[name] = version
-			}
-			return pkgs
-		},
-		GuessRegexps: pythonGuessRegexps,
-		Guess:        guess,
-		InstallReplitNixSystemDependencies: func(ctx context.Context, pkgs []api.PkgName) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "python.InstallReplitNixSystemDependencies")
-			defer span.Finish()
-			ops := []nix.NixEditorOp{}
-			for _, pkg := range pkgs {
-				deps := nix.PythonNixDeps(string(pkg))
-				ops = append(ops, nix.ReplitNixAddToNixEditorOps(deps)...)
-			}
-
-			// Ignore the error here, because if we can't read the specfile,
-			// we still want to add the deps from above at least.
-			specfilePkgs, _ := listPoetrySpecfile(true)
-			for pkg := range specfilePkgs {
-				deps := nix.PythonNixDeps(string(pkg))
-				ops = append(ops, nix.ReplitNixAddToNixEditorOps(deps)...)
-			}
-			nix.RunNixEditorOps(ops)
-		},
-	}
-}
-
 var pythonGuessRegexps = util.Regexps([]string{
 	// The (?:.|\\\n) subexpression allows us to
 	// match match multiple lines if
@@ -414,262 +204,12 @@ var pythonGuessRegexps = util.Regexps([]string{
 	`import ((?:.|\\\n)*)`,
 })
 
-// makePythonPipBackend returns a backend for invoking poetry, given an arg0 for invoking Python
-// (either a full path or just a name like "python3") to use when invoking Python.
-func makePythonPipBackend(python string) api.LanguageBackend {
-	var pipFlags []PipFlag
-
-	b := api.LanguageBackend{
-		Name:                 "python3-pip",
-		Specfile:             "requirements.txt",
-		IsAvailable:          pipIsAvailable,
-		Alias:                "python-python3-pip",
-		FilenamePatterns:     []string{"*.py"},
-		Quirks:               api.QuirksAddRemoveAlsoInstalls | api.QuirksNotReproducible,
-		NormalizePackageArgs: normalizePackageArgs,
-		NormalizePackageName: normalizePackageName,
-		GetPackageDir: func() string {
-			// Check if we're already inside an activated
-			// virtualenv. If so, just use it.
-			if venv := os.Getenv("VIRTUAL_ENV"); venv != "" {
-				return venv
-			}
-
-			// Take PYTHONUSERBASE into consideration, if set
-			if userbase := os.Getenv("PYTHONUSERBASE"); userbase != "" {
-				return userbase
-			}
-
-			if outputB, err := util.GetCmdOutputFallible([]string{
-				"python",
-				"-c", "import site; print(site.USER_BASE)",
-			}); err == nil {
-				return string(outputB)
-			}
-
-			return ""
-		},
-		SortPackages: pkg.SortPrefixSuffix(normalizePackageName),
-
-		Search: searchPypi,
-		Info:   info,
-		Add: func(ctx context.Context, pkgs map[api.PkgName]api.PkgSpec, projectName string) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "pip install")
-			defer span.Finish()
-
-			cmd := []string{"pip", "install"}
-			for _, flag := range pipFlags {
-				cmd = append(cmd, string(flag))
-			}
-			for name, spec := range pkgs {
-				name := string(name)
-				spec := string(spec)
-				if found, ok := moduleToPypiPackageAliases[name]; ok {
-					delete(pkgs, api.PkgName(name))
-					name = found
-					pkgs[api.PkgName(name)] = api.PkgSpec(spec)
-				}
-
-				cmd = append(cmd, name+" "+spec)
-			}
-			// Run install
-			util.RunCmd(cmd)
-			// Determine what was actually installed
-			outputB, err := util.GetCmdOutputFallible([]string{
-				"pip", "freeze",
-			})
-			if err != nil {
-				util.DieSubprocess("failed to run freeze: %s", err.Error())
-			}
-
-			// As we walk through the output of pip freeze,
-			// compare the package metadata name to the normalized
-			// pkgs that we are trying to install, to see which we
-			// want to track in `requirements.txt`.
-			normalizedPkgs := make(map[api.PkgName]api.PkgName)
-			for name := range pkgs {
-				normalizedPkgs[normalizePackageName(name)] = name
-			}
-
-			var toAppend []string
-			for _, canonicalSpec := range strings.Split(string(outputB), "\n") {
-				var name api.PkgName
-				matches := matchPackageAndSpec.FindSubmatch(([]byte)(canonicalSpec))
-				if len(matches) > 0 {
-					name = normalizePackageName(api.PkgName(string(matches[1])))
-					if rawName, ok := normalizedPkgs[name]; ok {
-						// We've meticulously maintained the pkgspec from the CLI args, if specified,
-						// so we don't clobber it with pip freeze's output of "==="
-						name := string(matches[1])
-						userArgSpec := string(pkgs[rawName])
-						toAppend = append(toAppend, name+userArgSpec)
-					}
-				}
-			}
-
-			handle, err := os.OpenFile("requirements.txt", os.O_APPEND|os.O_CREATE|os.O_RDWR, 0o644)
-			if err != nil {
-				util.DieIO("Unable to open requirements.txt for writing: %s", err)
-			}
-			defer handle.Close()
-
-			// Probe handle to determine if the last character is a newline
-			var hasTrailingNewline bool
-			fileInfo, err := handle.Stat()
-			if err != nil {
-				util.DieIO("Error getting file info: %s", err)
-			}
-			if fileInfo.Size() > 0 {
-				var lastChar = make([]byte, 1)
-				_, err := handle.ReadAt(lastChar, fileInfo.Size()-1)
-				if err != nil {
-					util.DieIO("Error reading last character: %s", err)
-				}
-				hasTrailingNewline = (lastChar[0] == '\n')
-			}
-			// Maintain existing formatting style.
-			// If the user has a trailing newline, keep it.
-			// If the user has no trailing newline, don't add one.
-			var leadingNewline, trailingNewline string
-			if hasTrailingNewline {
-				leadingNewline = ""
-				trailingNewline = "\n"
-			} else {
-				leadingNewline = "\n"
-				trailingNewline = ""
-			}
-			for _, line := range toAppend {
-				if _, err := handle.WriteString(leadingNewline + line + trailingNewline); err != nil {
-					util.DieIO("Error writing to requirements.txt: %s", err)
-				}
-			}
-		},
-		Remove: func(ctx context.Context, pkgs map[api.PkgName]bool) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "pip uninstall")
-			defer span.Finish()
-
-			cmd := []string{"pip", "uninstall", "--yes"}
-			for name := range pkgs {
-				cmd = append(cmd, string(name))
-			}
-			util.RunCmd(cmd)
-			err := RemoveFromRequirementsTxt("requirements.txt", pkgs)
-			if err != nil {
-				util.DieIO("%s", err.Error())
-			}
-		},
-		Install: func(ctx context.Context) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "pip install")
-			defer span.Finish()
-
-			util.RunCmd([]string{"pip", "install", "-r", "requirements.txt"})
-		},
-		ListSpecfile: func(mergeAllGroups bool) map[api.PkgName]api.PkgSpec {
-			flags, pkgs, err := ListRequirementsTxt("requirements.txt")
-			if err != nil {
-				util.DieIO("%s", err.Error())
-			}
-
-			// Stash the seen flags into a module global.
-			// This isn't great, but the expectation is that ListSpecfile
-			// is called before we run `Add`.
-			pipFlags = flags
-
-			// NB: We rely on requirements.txt being populated with the
-			// Python package _metadata_ name, not the PEP-503/PEP-508
-			// normalized version.
-			return pkgs
-		},
-		GuessRegexps: pythonGuessRegexps,
-		Guess:        guess,
-		InstallReplitNixSystemDependencies: func(ctx context.Context, pkgs []api.PkgName) {
-			//nolint:ineffassign,wastedassign,staticcheck
-			span, ctx := tracer.StartSpanFromContext(ctx, "python.InstallReplitNixSystemDependencies")
-			defer span.Finish()
-			ops := []nix.NixEditorOp{}
-			for _, pkg := range pkgs {
-				deps := nix.PythonNixDeps(string(pkg))
-				ops = append(ops, nix.ReplitNixAddToNixEditorOps(deps)...)
-			}
-
-			// Ignore the error here, because if we can't read the specfile,
-			// we still want to add the deps from above at least.
-			_, specfilePkgs, _ := ListRequirementsTxt("requirements.txt")
-			for pkg := range specfilePkgs {
-				deps := nix.PythonNixDeps(string(pkg))
-				ops = append(ops, nix.ReplitNixAddToNixEditorOps(deps)...)
-			}
-			nix.RunNixEditorOps(ops)
-		},
-	}
-
-	return b
-}
-
 func readPyproject() (*pyprojectTOML, error) {
 	var cfg pyprojectTOML
 	if _, err := toml.DecodeFile("pyproject.toml", &cfg); err != nil {
 		return nil, err
 	}
 	return &cfg, nil
-}
-
-func verifyPoetrySpecfile(path string) (bool, error) {
-	cfg, err := readPyproject()
-	if err != nil {
-		return false, err
-	}
-
-	return cfg.Tool.Poetry != nil, nil
-}
-
-func listPoetrySpecfile(mergeAllGroups bool) (map[api.PkgName]api.PkgSpec, error) {
-	cfg, err := readPyproject()
-	if err != nil {
-		return nil, err
-	}
-	pkgs := map[api.PkgName]api.PkgSpec{}
-	if cfg.Tool.Poetry == nil {
-		return pkgs, nil
-	}
-	for nameStr, spec := range cfg.Tool.Poetry.Dependencies {
-		if nameStr == "python" {
-			continue
-		}
-
-		specStr := normalizeSpec(spec)
-		if specStr == "" {
-			continue
-		}
-		pkgs[api.PkgName(nameStr)] = api.PkgSpec(specStr)
-	}
-	for nameStr, spec := range cfg.Tool.Poetry.DevDependencies {
-		if nameStr == "python" {
-			continue
-		}
-
-		specStr := normalizeSpec(spec)
-		if specStr == "" {
-			continue
-		}
-		pkgs[api.PkgName(nameStr)] = api.PkgSpec(specStr)
-	}
-	if mergeAllGroups && cfg.Tool.Poetry.Group != nil {
-		for _, group := range cfg.Tool.Poetry.Group {
-			for nameStr, spec := range group.Dependencies {
-				specStr := normalizeSpec(spec)
-				if specStr == "" {
-					continue
-				}
-				pkgs[api.PkgName(nameStr)] = api.PkgSpec(specStr)
-			}
-		}
-	}
-
-	return pkgs, nil
 }
 
 // getPython3 returns either "python3" or the value of the UPM_PYTHON3
